BAU: extract prompts and rename basic page to stub page
