+++ prompts/README.md
@@ -1,0 +1,1 @@
+Prompts for use in plopfile.
