+++ prompts/question-type-prompt.js
@@ -1,0 +1,6 @@
+export const questionTypePrompt = {
+  type: 'list',
+  name: 'questionType',
+  message: 'What \x1b[33mtype\x1b[0m of question are you adding?',
+  choices: ['text', 'text-area', 'radio-button', 'checkbox', 'number']
+}
