+++ templates/stub-page/index.js.hbs
